refac: Update test fixtures and clean up unused code in measurement tests
