+++ source/geh_calculated_measurements/tests/electrical_heating/job_tests/conftest.py
@@ -1,47 +1,23 @@
 import pytest
 from geh_common.pyspark.read_csv import read_csv_path
-from geh_common.testing.delta_lake.delta_lake_operations import create_database, create_table
 from pyspark.sql import SparkSession
 
-from geh_calculated_measurements.common.domain import CalculatedMeasurements, CurrentMeasurements
-from geh_calculated_measurements.common.infrastructure import CalculatedMeasurementsInternalDatabaseDefinition
+from geh_calculated_measurements.common.domain import CurrentMeasurements
 from geh_calculated_measurements.common.infrastructure.current_measurements.database_definitions import (
     MeasurementsGoldDatabaseDefinition,
 )
 from tests.electrical_heating.job_tests import get_test_files_folder_path
 
 
-# TODO BJM: Rely on a fixture to set up these shared tables?
 @pytest.fixture(scope="session")
-def gold_table_seeded(spark: SparkSession) -> None:
-    create_database(spark, MeasurementsGoldDatabaseDefinition.DATABASE_NAME)
-
-    create_table(
-        spark,
-        database_name=MeasurementsGoldDatabaseDefinition.DATABASE_NAME,
-        table_name=MeasurementsGoldDatabaseDefinition.CURRENT_MEASUREMENTS,
-        schema=CurrentMeasurements.schema,
-        table_location=f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}/{MeasurementsGoldDatabaseDefinition.CURRENT_MEASUREMENTS}",
-    )
-
+def gold_table_seeded(
+    spark: SparkSession,
+    external_dataproducts_created: None,  # Used implicitly
+) -> None:
     file_name = f"{get_test_files_folder_path()}/{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}-{MeasurementsGoldDatabaseDefinition.CURRENT_MEASUREMENTS}.csv"
     time_series_points = read_csv_path(spark, file_name, CurrentMeasurements.schema)
-    time_series_points.write.option("overwriteSchema", "true").saveAsTable(
+    time_series_points.write.saveAsTable(
         f"{MeasurementsGoldDatabaseDefinition.DATABASE_NAME}.{MeasurementsGoldDatabaseDefinition.CURRENT_MEASUREMENTS}",
         format="delta",
         mode="overwrite",
-    )
-
-
-# TODO BJM: Rely on a fixture to set up these shared tables?
-@pytest.fixture(scope="session")
-def calculated_measurements_table_created(spark: SparkSession) -> None:
-    create_database(spark, CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME)
-
-    create_table(
-        spark,
-        database_name=CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME,
-        table_name=CalculatedMeasurementsInternalDatabaseDefinition.MEASUREMENTS_TABLE_NAME,
-        schema=CalculatedMeasurements.schema,
-        table_location=f"{CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME}/{CalculatedMeasurementsInternalDatabaseDefinition.MEASUREMENTS_TABLE_NAME}",
     )
