+++ source/geh_calculated_measurements/tests/electrical_heating/job_tests/test_electrical_heating_job.py
@@ -17,7 +17,7 @@ from tests.electrical_heating.job_tests import get_test_files_folder_path
 def test_execute(
     spark: SparkSession,
     gold_table_seeded: Any,  # Used implicitly
-    calculated_measurements_table_created: Any,  # Used implicitly
+    migrations_executed: None,  # Used implicitly
     dummy_logging: Any,  # Used implicitly
     monkeypatch,
 ) -> None:
