+++ source/geh_calculated_measurements/tests/integration/database_migrations/test_migrate.py
@@ -7,14 +7,12 @@ from geh_calculated_measurements.testing.utilities.create_azure_log_query_runner
     LogsQueryStatus,
     create_azure_log_query_runner,
 )
-from tests import ensure_calculated_measurements_databases_exist
 
 
 def test__when_running_migrate__then_log_is_produced(spark: SparkSession, monkeypatch: pytest.MonkeyPatch):
     # Arrange
     azure_query_runnner = create_azure_log_query_runner(monkeypatch)
     timeout_minutes = 15
-    ensure_calculated_measurements_databases_exist(spark)
     monkeypatch.setattr(
         migrations_runner, "_migrate", lambda name, subs: None
     )  # Mock this function to avoid actual migration
