+++ source/geh_calculated_measurements/tests/missing_measurements_log/job_tests/test_missing_measurements_log_job.py
@@ -1,8 +1,6 @@
 import os
 import uuid
 
-from pyspark.sql import SparkSession
-
 from geh_calculated_measurements.missing_measurements_log.entry_point import execute
 from tests import create_job_environment_variables
 
@@ -20,10 +18,9 @@ def _create_job_arguments(orchestration_instance_id: uuid.UUID) -> list[str]:
 
 
 def test_execute(
-    spark: SparkSession,
     monkeypatch,
     dummy_logging: None,  # Used implicitly
-    input_tables_created: None,
+    external_dataproducts_created: None,  # Used implicitly
 ) -> None:
     # Arrange
     orchestration_instance_id = uuid.uuid4()
