+++ source/geh_calculated_measurements/tests/subsystem_tests/net_consumption_group_6/seed_table.py
@@ -2,7 +2,7 @@ from datetime import datetime, timezone
 
 from geh_calculated_measurements.testing.utilities.job_tester import JobTestFixture
 
-# TODO BJM: Use values from production code
+# TODO Henrik: Use values from production code
 database = "electricity_market_measurements_input"
 parent_table = "net_consumption_group_6_consumption_metering_point_periods_v1"
 child_table = "net_consumption_group_6_child_metering_point_periods_v1"
